--- a/src/SharpNeat/Neat/Genome/INeatGenomeBuilder.cs
+++ b/src/SharpNeat/Neat/Genome/INeatGenomeBuilder.cs
@@ -1,5 +1,6 @@
 ﻿// This file is part of SharpNEAT; Copyright Colin D. Green.
 // See LICENSE.txt for details.
+using System.Numerics;
 
 namespace SharpNeat.Neat.Genome;
 
@@ -8,7 +9,7 @@ namespace SharpNeat.Neat.Genome;
 /// </summary>
 /// <typeparam name="TScalar">Neural net connection weight and signal data type.</typeparam>
 public interface INeatGenomeBuilder<TScalar>
-    where TScalar : unmanaged
+    where TScalar : unmanaged, IBinaryFloatingPointIeee754<TScalar>
 {
     /// <summary>
     /// Create a NeatGenome with the given meta data and connection genes.