--- a/src/SharpNeat/Neat/Genome/NeatGenomeAssertions.cs
+++ b/src/SharpNeat/Neat/Genome/NeatGenomeAssertions.cs
@@ -1,6 +1,7 @@
 ﻿// This file is part of SharpNEAT; Copyright Colin D. Green.
 // See LICENSE.txt for details.
 using System.Diagnostics;
+using System.Numerics;
 using Redzen.Sorting;
 using SharpNeat.Graphs.Acyclic;
 
@@ -12,7 +13,7 @@ namespace SharpNeat.Neat.Genome;
 /// </summary>
 /// <typeparam name="TScalar">Neural net connection weight and signal data type.</typeparam>
 internal static class NeatGenomeAssertions<TScalar>
-    where TScalar : unmanaged
+    where TScalar : unmanaged, IBinaryFloatingPointIeee754<TScalar>
 {
     #region Public Static Methods