--- a/src/SharpNeat/Neat/Reproduction/Asexual/Strategies/AddNodeStrategy.cs
+++ b/src/SharpNeat/Neat/Reproduction/Asexual/Strategies/AddNodeStrategy.cs
@@ -14,7 +14,7 @@ namespace SharpNeat.Neat.Reproduction.Asexual.Strategies;
 /// A → B with A → C → B, where A and B are the existing nodes, and C is the new node.
 /// </remarks>
 public sealed class AddNodeStrategy<TScalar> : IAsexualReproductionStrategy<TScalar>
-    where TScalar : unmanaged, INumberBase<TScalar>
+    where TScalar : unmanaged, IBinaryFloatingPointIeee754<TScalar>
 {
     readonly MetaNeatGenome<TScalar> _metaNeatGenome;
     readonly INeatGenomeBuilder<TScalar> _genomeBuilder;