--- a/src/SharpNeat/Neat/Speciation/Species.cs
+++ b/src/SharpNeat/Neat/Speciation/Species.cs
@@ -1,5 +1,7 @@
 ﻿// This file is part of SharpNEAT; Copyright Colin D. Green.
 // See LICENSE.txt for details.
+using System.Numerics;
+
 namespace SharpNeat.Neat.Speciation;
 
 /// <summary>
@@ -7,7 +9,7 @@ namespace SharpNeat.Neat.Speciation;
 /// </summary>
 /// <typeparam name="TScalar">Neural net connection weight and signal data type.</typeparam>
 public class Species<TScalar>
-    where TScalar : unmanaged
+    where TScalar : unmanaged, IBinaryFloatingPointIeee754<TScalar>
 {
     /// <summary>
     /// Species ID.