--- a/src/SharpNeat/NeuralNets/ActivationFunctions/ArcSinH.cs
+++ b/src/SharpNeat/NeuralNets/ActivationFunctions/ArcSinH.cs
@@ -3,7 +3,7 @@
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-namespace SharpNeat.NeuralNets.Double.ActivationFunctions;
+namespace SharpNeat.NeuralNets.ActivationFunctions;
 
 /// <summary>
 /// The ArcSinH function (inverse hyperbolic sine function).