--- a/src/SharpNeat/NeuralNets/ActivationFunctions/ArcTan.cs
+++ b/src/SharpNeat/NeuralNets/ActivationFunctions/ArcTan.cs
@@ -3,7 +3,7 @@
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-namespace SharpNeat.NeuralNets.Double.ActivationFunctions;
+namespace SharpNeat.NeuralNets.ActivationFunctions;
 
 /// <summary>
 /// The ArcTan function (inverse tangent function).