… src/Benchmarks/SharpNeat.Benchmarks/Neat/DistanceMetrics/DistanceMetricUtilsBenchmarks.cs
… src/Benchmarks/SharpNeat.Benchmarks/Neat/DistanceMetrics/EuclideanDistanceMetricBenchmarks.cs
… src/Benchmarks/SharpNeat.Benchmarks/Neat/DistanceMetrics/ManhattanDistanceMetricBenchmarks.cs
… src/Benchmarks/SharpNeat.Benchmarks/NeuralNets/ActivationFunctions/ActivationFunctionsBenchmarks.cs
… src/Benchmarks/SharpNeat.Benchmarks/NeuralNets/ActivationFunctions/Vectorized/ActivationFunctionsBenchmarks.cs
… src/Benchmarks/SharpNeat.Benchmarks/NeuralNets/Double/ActivationFunctions/ActivationFunctionsBenchmarks.cs
… src/Benchmarks/SharpNeat.Benchmarks/NeuralNets/Double/ActivationFunctions/Vectorized/ActivationFunctionsBenchmarks.cs
… src/Benchmarks/SharpNeat.Benchmarks/Program.cs
… src/SharpNeat/NeuralNets/ActivationFunctions/Cppn/BipolarGaussian.cs
… src/SharpNeat/NeuralNets/ActivationFunctions/Cppn/BipolarSigmoid.cs
… src/SharpNeat/NeuralNets/ActivationFunctions/Cppn/Gaussian.cs
… src/SharpNeat/NeuralNets/ActivationFunctions/Cppn/Linear.cs
… src/SharpNeat/NeuralNets/ActivationFunctions/Cppn/Sine.cs
… src/SharpNeat/NeuralNets/ActivationFunctions/LeakyReLU.cs
… src/SharpNeat/NeuralNets/ActivationFunctions/LeakyReLUShifted.cs
… src/SharpNeat/NeuralNets/ActivationFunctions/Logistic.cs
… src/SharpNeat/NeuralNets/ActivationFunctions/LogisticApproximantSteep.cs
… src/SharpNeat/NeuralNets/ActivationFunctions/LogisticSteep.cs
… src/SharpNeat/NeuralNets/ActivationFunctions/MaxMinusOne.cs
… src/SharpNeat/NeuralNets/ActivationFunctions/NullFn.cs
… src/SharpNeat/NeuralNets/ActivationFunctions/PolynomialApproximantSteep.cs
… src/SharpNeat/NeuralNets/ActivationFunctions/QuadraticSigmoid.cs
… src/SharpNeat/NeuralNets/ActivationFunctions/ReLU.cs
… src/SharpNeat/NeuralNets/ActivationFunctions/SReLU.cs
… src/SharpNeat/NeuralNets/ActivationFunctions/SReLUShifted.cs
… src/SharpNeat/NeuralNets/ActivationFunctions/ScaledELU.cs
… src/SharpNeat/NeuralNets/ActivationFunctions/SoftSignSteep.cs
… src/SharpNeat/NeuralNets/ActivationFunctions/TanH.cs
… src/SharpNeat/NeuralNets/ActivationFunctions/Vectorized/LeakyReLU.cs
… src/SharpNeat/NeuralNets/ActivationFunctions/Vectorized/LeakyReLUShifted.cs