Activation functions made generic.
Some vectorized fn optimizations (LeakyReLU, LeakyReLUShifted.)

# Benchmarks

### Before (non-generic)

scalar-double
| Method                     | Mean       | Error    | StdDev   |
|--------------------------- |-----------:|---------:|---------:|
| ArcSinH                    | 6,908.5 us | 41.65 us | 38.96 us |
| ArcTan                     | 5,656.0 us | 27.10 us | 22.63 us |
| LeakyReLU                  |   884.4 us |  3.43 us |  3.04 us |
| LeakyReLUShifted           |   753.9 us |  3.34 us |  3.13 us |
| Logistic                   | 4,426.9 us | 13.62 us | 12.74 us |
| LogisticSteep              | 4,487.0 us | 16.41 us | 14.55 us |
| MaxMinusOne                |   582.9 us |  3.31 us |  2.93 us |
| NullFn                     |   453.6 us |  1.15 us |  1.07 us |
| PolynomialApproximantSteep | 1,723.4 us |  3.63 us |  3.39 us |
| QuadraticSigmoid           | 1,498.3 us |  5.12 us |  4.79 us |
| ReLU                       |   889.7 us |  3.57 us |  3.34 us |
| ScaledELU                  | 2,844.1 us |  7.08 us |  6.28 us |
| SoftSignSteep              | 1,002.3 us |  1.19 us |  0.93 us |
| SReLU                      | 1,041.4 us |  3.26 us |  2.72 us |
| SReLUShifted               | 1,118.7 us |  3.98 us |  3.53 us |
| TanH                       | 8,645.5 us | 21.52 us | 20.13 us |

vectorized-double
| Method                     | Mean       | Error   | StdDev  |
|--------------------------- |-----------:|--------:|--------:|
| LeakyReLU                  |   311.3 us | 1.79 us | 1.67 us |
| LeakyReLUShifted           |   342.4 us | 1.22 us | 1.14 us |
| MaxMinusOne                |   280.0 us | 0.79 us | 0.66 us |
| PolynomialApproximantSteep |   542.3 us | 1.30 us | 1.08 us |
| QuadraticSigmoid           | 1,966.1 us | 2.96 us | 2.77 us |
| ReLU                       |   125.4 us | 0.38 us | 0.34 us |
| SoftSignSteep              |   275.6 us | 1.06 us | 0.99 us |
| SReLU                      |   276.5 us | 0.69 us | 0.54 us |
| SReLUShifted               |   347.6 us | 0.72 us | 0.64 us |


### After (generic + some optimisations)

scalar-generic-double
| Method                     | Mean       | Error    | StdDev   |
|--------------------------- |-----------:|---------:|---------:|
| ArcSinH                    | 6,817.0 us | 39.17 us | 36.64 us |
| ArcTan                     | 5,634.8 us | 26.23 us | 24.54 us |
| LeakyReLU                  |   848.0 us |  1.49 us |  1.32 us |
| LeakyReLUShifted           |   656.5 us |  1.95 us |  1.83 us |
| Logistic                   | 4,821.9 us | 30.18 us | 28.23 us |
| LogisticSteep              | 4,734.3 us |  7.26 us |  6.07 us |
| MaxMinusOne                |   539.1 us |  1.76 us |  1.56 us |
| NullFn                     |   452.3 us |  1.43 us |  1.34 us |
| PolynomialApproximantSteep | 1,748.2 us |  4.95 us |  4.63 us |
| QuadraticSigmoid           | 1,620.8 us |  6.95 us |  6.50 us |
| ReLU                       |   454.6 us |  1.73 us |  1.61 us |
| ScaledELU                  | 2,861.7 us | 14.97 us | 12.50 us |
| SoftSignSteep              |   999.4 us |  2.72 us |  2.54 us |
| SReLU                      | 1,184.2 us |  3.32 us |  2.94 us |
| SReLUShifted               | 1,118.7 us | 15.55 us | 22.31 us |
| TanH                       | 9,028.3 us | 12.68 us | 10.59 us |


vectorized-generic-double
| Method                     | Mean       | Error    | StdDev   |
|--------------------------- |-----------:|---------:|---------:|
| LeakyReLU                  |   125.6 us |  0.19 us |  0.16 us |
| LeakyReLUShifted           |   129.1 us |  0.58 us |  0.49 us |
| MaxMinusOne                |   191.7 us |  0.62 us |  0.52 us |
| PolynomialApproximantSteep |   537.2 us |  2.30 us |  2.15 us |
| QuadraticSigmoid           | 1,920.2 us |  2.34 us |  1.96 us |
| ReLU                       |   125.7 us |  0.55 us |  0.51 us |
| SoftSignSteep              |   265.9 us |  0.41 us |  0.36 us |
| SReLU                      |   289.7 us |  1.15 us |  1.02 us |
| SReLUShifted               |   318.2 us |  1.26 us |  1.17 us |